Error Fix. User Error Checking First Pass.
"Fixed" a long standing error wherein SpellEntryForm design was creating
an error on first load in VS. Fixed by unlinking SpellEntryForm.cs and
making it exist in SpellEntry project (so basically the files were moved
to SpellEntry from Spells.

Created a first pass implementation for Error Notification.
ID, Name, Casting Time, Duration, Range, and Duration labels now turn
RED if their values are invalid.
The above as well as Material Cost and Higher Levels fields will alert
the user if they try to save a spell with values entered in the fields
when they are disabled.