Removed Close and Import buttons
Apparently I deleted more stuff? This commit should have happened a
month ago when I actually made these changes....
- Removed Close button because it was pointless. RIP.
- Removed Import button because that functionality will be handled by
the Add/Edit Spells button which will open the XML Writer. XML Writer
will need to be modified to write to the database. Previously the import
button called addRowsToTable(..), which writes to the SQLite database.
This method still exists but is not currently being called by anything.
- loadSelectedDatabase() removed because it is now linked directly to
the CoBx selected index changing, so that functionality exists in that
event listener now